apply new CEKClient
