+++ clova-extensions-kit-csharp-azure-functions/clova-extensions-kit-csharp-azure-functions.csproj
@@ -2,7 +2,7 @@
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
     <AzureFunctionsVersion>v2</AzureFunctionsVersion>
-    <RootNamespace>clova_extensions_kit_csharp_azure_functions</RootNamespace>
+    <RootNamespace>LineDC.CEK</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.DurableTask" Version="1.8.0" />
